% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_absnorm.R
\name{absnorm}
\alias{absnorm}
\alias{r_absnorm}
\alias{p_absnorm}
\alias{q_absnorm}
\title{The Absolute Normal Distribution}
\usage{
r_absnorm(n, mean = 0, sd = 1)

p_absnorm(q, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)

q_absnorm(p, mean = 0, sd = 1)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
    \eqn{P[X \le x]} otherwise, \eqn{P[X > x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{vector of probabilities.}
}
\description{
distribution function, quantile function and
random generation for the absolute value normal distribution
with mean equal to [mean] and standard deviation equal to [sd].
}
