% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_intervals.R
\name{are_disjoint_intervals}
\alias{are_disjoint_intervals}
\title{Check if Two Intervals are Disjoint}
\usage{
are_disjoint_intervals(int_1, int_2)
}
\arguments{
\item{int_1}{double(2) \cr
An interval to intersect with [int_2]}

\item{int_2}{double(2) \cr
An interval to intersect with [int_1]}
}
\description{
Check if Two Intervals are Disjoint
}
