% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defense.R
\name{check_beta}
\alias{check_beta}
\title{Check 'beta' Argument}
\usage{
check_beta(beta, conf_level)
}
\arguments{
\item{beta}{double in [0, 1 - conf_level] \cr
For hybrid intervals, degree of mixing between conditional and unconditional
intervals. At 0, hybrid intervals are identical to conditional intervals.}

\item{conf_level}{double in (0,1) \cr
The desired confidence level of the resulting interval. Note that
conditional and hybrid confidence intervals don't have nominal
coverage in the conventional sense.}
}
\description{
Check 'beta' Argument
}
