% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defense.R
\name{check_conf_level}
\alias{check_conf_level}
\title{Check 'conf_level' Argument}
\usage{
check_conf_level(conf_level)
}
\arguments{
\item{conf_level}{double in (0,1) \cr
The desired confidence level of the resulting interval. Note that
conditional and hybrid confidence intervals don't have nominal
coverage in the conventional sense.}
}
\description{
Check 'conf_level' Argument
}
