% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defense.R
\name{check_estimates}
\alias{check_estimates}
\title{Check 'estimates' Argument}
\usage{
check_estimates(estimates)
}
\arguments{
\item{estimates}{numeric vector \cr
A vector of estimates.}
}
\description{
Check 'estimates' Argument
}
