% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defense.R
\name{check_st_errs}
\alias{check_st_errs}
\title{Check 'st_errs' Argument}
\usage{
check_st_errs(st_errs, estimates)
}
\arguments{
\item{st_errs}{numeric vector \cr
A vector of estimated standard errors for each estimate.}

\item{estimates}{numeric vector \cr
A vector of estimates. Will return a confidence interval for the maximum
estimate.}
}
\description{
Check 'st_errs' Argument
}
