% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_docs.R
\name{ci_docs}
\alias{ci_docs}
\title{Confidence Interval Documentation}
\arguments{
\item{estimates}{numeric vector \cr
A vector of estimates.}

\item{st_errs}{numeric vector \cr
A vector of estimated standard errors for each estimate.}

\item{k}{integer \cr
The rank of the estimate of interest. k = 1 returns the confidence interval
for the highest estimate, and k = length(estimates) returns the
confidence interval for the lowest estimate.}

\item{conf_level}{double in (0,1) \cr
The desired confidence level of the resulting interval. Note that
conditional and hybrid confidence intervals don't have nominal
coverage in the conventional sense.}

\item{beta}{double in [0, 1 - conf_level] \cr
For hybrid intervals, degree of mixing between conditional and unconditional
intervals. At 0, hybrid intervals are identical to conditional intervals.}
}
\description{
Confidence Interval Documentation
}
