% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int_hybrid_equal_tailed.R
\name{ci_hybrid_equal_tailed}
\alias{ci_hybrid_equal_tailed}
\title{Equal-Tailed Hybrid Confidence Interval}
\usage{
ci_hybrid_equal_tailed(
  estimates,
  st_errs,
  conf_level = 0.95,
  beta = (1 - conf_level)/10
)
}
\arguments{
\item{estimates}{numeric vector \cr
A vector of estimates. Will return a confidence interval for the maximum
estimate.}

\item{st_errs}{numeric vector \cr
A vector of estimated standard errors for each estimate.}

\item{conf_level}{double in (0,1) \cr
The desired confidence level of the resulting interval. Note that
conditional and hybrid confidence intervals don't have nominal
coverage in the conventional sense.}

\item{beta}{double in [0, 1 - conf_level] \cr
For hybrid intervals, degree of mixing between conditional and unconditional
intervals. At 0, hybrid intervals are identical to conditional intervals.}
}
\description{
Equal-Tailed Hybrid Confidence Interval
}
