% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int_proportion.R
\name{ci_proportion}
\alias{ci_proportion}
\title{Confidence Interval for a Proportion}
\usage{
ci_proportion(y, n = length(y), conf_level = 0.95)
}
\arguments{
\item{y}{integer | vector \cr
Either the number of successes, or a vector of successes and failures}

\item{n}{integer \cr
If y is the number of successes, the number of trials.}

\item{conf_level}{double in (0,1) \cr
The desired confidence level of the resulting interval. Note that
conditional and hybrid confidence intervals don't have nominal
coverage in the conventional sense.}
}
\description{
Confidence Interval for a Proportion
}
