% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int_unconditional.R
\name{ci_unconditional}
\alias{ci_unconditional}
\title{Unconditional Confidence Interval Using Order Statistics}
\usage{
ci_unconditional(estimates, st_errs, conf_level = 0.95)
}
\arguments{
\item{estimates}{numeric vector \cr
A vector of estimates. Will return a confidence interval for the maximum
estimate.}

\item{st_errs}{numeric vector \cr
A vector of estimated standard errors for each estimate.}

\item{conf_level}{double in (0,1) \cr
The desired confidence level of the resulting interval. Note that
conditional and hybrid confidence intervals don't have nominal
coverage in the conventional sense.}
}
\description{
Unconditional Confidence Interval Using Order Statistics
}
