% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncation_points.R
\name{conditional_trunc_lo}
\alias{conditional_trunc_lo}
\title{Lower Truncation Point}
\usage{
conditional_trunc_lo(estimates)
}
\arguments{
\item{estimates}{numeric vector \cr
A vector of estimates. Will return a confidence interval for the maximum
estimate.}
}
\description{
Lower Truncation Point
}
