% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncation_points.R
\name{conditional_trunc_points}
\alias{conditional_trunc_points}
\title{Truncation Points}
\usage{
conditional_trunc_points(estimates, k)
}
\arguments{
\item{estimates}{numeric vector \cr
A vector of estimates.}

\item{k}{integer \cr
The rank of the estimate of interest. k = 1 returns the confidence interval
for the highest estimate, and k = length(estimates) returns the
confidence interval for the lowest estimate.}
}
\description{
Truncation Points
}
