% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lagmatrix.R
\name{create_lagmatrix}
\alias{create_lagmatrix}
\title{create_lagmatrix}
\usage{
create_lagmatrix(timeseries, k)
}
\value{
a matrix with k rows and n columns, where n is the length of the provided time series
}
\description{
This function creates the delay reconstruction, i.e. a matrix
where the first row contains y_1, ..., y_T (the given univariate time series)
the second row contains NA, y_2, ..., y_T .... etc
the k-th row contains NA, NA, ..., y_{k+1}, ..., y_T
}

