% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_tnorm_times_indicator.R
\name{e_truncnorm_times_indicator}
\alias{e_truncnorm_times_indicator}
\title{Expected Value of Truncated Normal Times Indicator}
\usage{
e_truncnorm_times_indicator(c_z, mu, sigma, trunc_lo = -Inf, trunc_hi = Inf)
}
\arguments{
\item{c_z}{interval for indicator}

\item{mu}{mean of the normal distribution pre-truncation}

\item{sigma}{standard deviation of the normal distribution pre-truncation}

\item{trunc_lo}{lower truncation point}

\item{trunc_hi}{upper truncation point}
}
\description{
E[z * 1(z in c_z)]
}
