% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_jump.R
\name{find_jump}
\alias{find_jump}
\title{Find The Jump in a Step Function}
\usage{
find_jump(.f, .bounds, .tol = 1e-07, ...)
}
\arguments{
\item{.f}{step function}

\item{.bounds}{endpoints within which the function steps}

\item{.tol}{the desired accuracy}

\item{...}{additional arguments to .f}
}
\value{
the input at which the function steps
}
\description{
Find The Jump in a Step Function
}
