% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_levydriven.R
\name{get_levydriven}
\alias{get_levydriven}
\title{Levy driven stochastic volatility model}
\usage{
get_levydriven()
}
\value{
The list contains rprior, dprior (generate and evaluate the density of prior distribution),
generate_randomness (generate data-generating variables), robservation (create synthetic
data sets), parameter_names (useful for plotting), thetadim (dimension of parameter),
ydim (dimension of observations), parameters (list of hyperparameters,
to be passed to rprior,dprior,robservation)
}
\description{
This function returns a list representing a Levy driven stochastic volatility model.
}

