% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_mgandk.R
\name{get_mgandk}
\alias{get_mgandk}
\title{multivariate G and k model}
\usage{
get_mgandk()
}
\value{
The list contains rprior, dprior (generate and evaluate the density of prior distribution),
generate_randomness (generate data-generating variables), robservation (create synthetic
data sets), parameter_names (useful for plotting), thetadim (dimension of parameter),
ydim (dimension of observations), parameters (list of hyperparameters,
to be passed to rprior,dprior,robservation)
}
\description{
This function returns
}
