% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_multivariatenormal.R
\name{get_multivariate_normal}
\alias{get_multivariate_normal}
\title{Multivariate Normal model}
\usage{
get_multivariate_normal(dimension)
}
\value{
The list contains rprior, dprior (generate and evaluate the density of prior distribution),
generate_randomness (generate data-generating variables), robservation (create synthetic
data sets), parameter_names (useful for plotting), thetadim (dimension of parameter),
ydim (dimension of observations), parameters (list of hyperparameters,
to be passed to rprior,dprior,robservation)
}
\description{
This function returns a list representing
a Normal location model, in a dimension specified by the user as an argument.
}

