% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_normal.R
\name{get_normal}
\alias{get_normal}
\title{Normal model}
\usage{
get_normal()
}
\value{
The list contains rprior, dprior (generate and evaluate the density of prior distribution),
generate_randomness (generate data-generating variables), robservation (create synthetic
data sets), parameter_names (useful for plotting), thetadim (dimension of parameter),
ydim (dimension of observations), parameters (list of hyperparameters,
to be passed to rprior,dprior,robservation)
}
\description{
This function returns a list representing
a Normal location model.
The prior is mu ~ Normal(mu_0, nu^{-1}), where nu is precision
and  tau ~ Gamma(alpha, beta), where beta is rate (1/scale).
 The likelihood is Y ~ Normal(mu, tau^2) where tau^2 is the variance.
}
