% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_ricker.R
\name{get_ricker}
\alias{get_ricker}
\title{Ricker model}
\usage{
get_ricker()
}
\value{
The list contains rprior, dprior (generate and evaluate the density of prior distribution),
generate_randomness (generate data-generating variables), robservation (create synthetic
data sets), parameter_names (useful for plotting), thetadim (dimension of parameter),
ydim (dimension of observations), parameters (list of hyperparameters,
to be passed to rprior,dprior,robservation)
}
\description{
This function returns a list representing
the Ricker model in
Wood (2010) Statistical inference for noisy nonlinear ecological dynamic systems
}
