% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_toggleswitch.R
\name{get_toggleswitch}
\alias{get_toggleswitch}
\title{Toggle switch model}
\usage{
get_toggleswitch()
}
\value{
The list contains rprior, dprior (generate and evaluate the density of prior distribution),
generate_randomness (generate data-generating variables), robservation (create synthetic
data sets), parameter_names (useful for plotting), thetadim (dimension of parameter),
ydim (dimension of observations), parameters (list of hyperparameters,
to be passed to rprior,dprior,robservation)
}
\description{
This function returns a list representing the toggle switch model
of Bonassi, F. V., West, M., et al. (2015).
Sequential Monte Carlo with adaptive weights for approximate Bayesian computation. Bayesian Analysis, 10(1):171-187.
}
