% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilbert_order.R
\name{hilbert_order}
\alias{hilbert_order}
\title{hilbert_order}
\usage{
hilbert_order(x)
}
\value{
a vector of index corresponding to the ordered samples
}
\description{
This function returns the "Hilbert order" of a sample of n
points of dimension d, stored in a matrix with d rows and n columns
where d is the dimension of each sample and n the number of samples.
The function essentially calls Hilbert_Sort_CGAL, from the CGAL library.
}
