% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_intervals.R
\name{intersect_intervals}
\alias{intersect_intervals}
\title{Return the Intersection of Two Intervals}
\usage{
intersect_intervals(int_1, int_2)
}
\arguments{
\item{int_1}{double(2) \cr
An interval to intersect with [int_2]}

\item{int_2}{double(2) \cr
An interval to intersect with [int_1]}
}
\description{
Return the Intersection of Two Intervals
}
