% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_intervals.R
\name{is_contained}
\alias{is_contained}
\title{Check if a Point/Interval is Contained in an Interval}
\usage{
is_contained(this, in_this)
}
\arguments{
\item{this}{double(2) \cr
A point or interval that may fall inside the interval [in_this]}

\item{in_this}{double(2) \cr
An interval that may contain the point/interval [this]}
}
\description{
Check if a Point/Interval is Contained in an Interval
}
