% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_max_absnorm.R
\name{max_absnorm}
\alias{max_absnorm}
\alias{r_max_absnorm}
\alias{p_max_absnorm}
\alias{q_max_absnorm}
\title{The Maximum Absolute Normal Distribution}
\usage{
r_max_absnorm(n, size, mean = 0, sd = 1)

p_max_absnorm(size, q, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)

q_max_absnorm(size, p, mean = 0, sd = 1)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{size}{number of absolute normal variables maximized over per
observation}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
    \eqn{P[X \le x]} otherwise, \eqn{P[X > x]}.}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{p}{vector of probabilities.}
}
\description{
distribution function, quantile function and
random generation for the maximum of [size] absolute value normal random
variables with mean equal to [mean] and standard deviation equal to [sd].
}
