% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_pz_transition.R
\name{pz_transition}
\alias{pz_transition}
\title{pz_transition}
\usage{
pz_transition(xparticles, alphas, time, parameters)
}
\description{
Solve PZ ODE for each particle, given each alpha, from time to time + 1,
and given the parameters (c, e, ml, mq).
}

