% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_experiment.R
\name{simulate_experiments}
\alias{simulate_experiments}
\alias{simulate_experiment}
\title{Simulate Experiment(s)}
\usage{
simulate_experiments(n_experiments, true_means, st_errs, n_obs = 100)

simulate_experiment(true_means, st_errs, n_obs = 100)
}
\arguments{
\item{n_experiments}{integer \cr
Number of experiments to simulate}

\item{true_means}{double vector \cr
Vector of true mean outcome for each treatment arm.}

\item{st_errs}{double vector \cr
Standard error of the sample mean in each treatment arm.}

\item{n_obs}{integer vector \cr
Number of observations in each treatment arm (default: 100).}
}
\value{
An experiment is a named list. simulate_experiment returns
a single list, while simulate_experiments returns a list of sublists where
each sublist is an experiment. An experiment has entries: \cr
\describe{
  \item{true_means}{The true mean outcome for each treatment arm.}
  \item{estimates}{The sample mean for each treatment arm.}
  \item{st_errs}{The true standard error of the sample means.}
  \item{st_errs_hat}{The estimated standard error of the sample means.}
}
}
\description{
simulate experiment(s) with many treatment arms, where
true treatment effects in each arm are given by [true_means],
the standard errors of the estimated treatment effects is given by
[st_errs], and the sample size in each treatment arm is given by [n_obs].
}
