% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_truncnorm.R
\name{truncnorm}
\alias{truncnorm}
\alias{p_truncnorm}
\alias{d_truncnorm}
\alias{q_truncnorm}
\alias{e_truncnorm}
\alias{r_truncnorm}
\title{Expected Value of the Truncated Normal Distribution}
\usage{
p_truncnorm(q, mean = 0, sd = 1, trunc_lo = -Inf, trunc_hi = Inf)

d_truncnorm(x, mean = 0, sd = 1, trunc_lo = -Inf, trunc_hi = Inf)

q_truncnorm(p, mean = 0, sd = 1, trunc_lo = -Inf, trunc_hi = Inf)

e_truncnorm(mean = 0, sd = 1, trunc_lo = -Inf, trunc_hi = Inf)

r_truncnorm(n = 1, mean = 0, sd = 1, trunc_lo = -Inf, trunc_hi = Inf)
}
\arguments{
\item{q}{vector of quantiles.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}

\item{trunc_lo}{lower truncation piont}

\item{trunc_hi}{upper truncation piont}

\item{x}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}
}
\description{
cumulative distribution, density, quantiles, and expected values
for the truncated normal distribution
with mean equal to [mean] and standard deviation equal to [sd],
truncated below at [trunc_lo] and above at [trunc_hi].
}
