% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variouswasserstein.R
\name{wasserstein_SAG}
\alias{wasserstein_SAG}
\title{#'@export
wasserstein <- function(p, qs, cost_matrix, epsilon, niterations){
  return(wasserstein_(p, qs, cost_matrix, epsilon, niterations))
}}
\usage{
wasserstein_SAG(p, q, cost_matrix, epsilon, niterations, stepsize)
}
\description{
#'@export
wasserstein <- function(p, qs, cost_matrix, epsilon, niterations){
  return(wasserstein_(p, qs, cost_matrix, epsilon, niterations))
}
}

