% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winference-package.R
\docType{package}
\name{winference-package}
\alias{winference}
\alias{winference-package}
\title{winference}
\description{
...
}
\details{
...
}
\author{
anonymous <anonymous@anonymous.com>
}
\keyword{package}

