% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int_conditional_equal_tailed.R
\name{ci_conditional_equal_tailed}
\alias{ci_conditional_equal_tailed}
\title{Conditional Confidence Interval Using Truncated Normality}
\usage{
ci_conditional_equal_tailed(estimates, st_errs, k = 1, conf_level = 0.95)
}
\arguments{
\item{estimates}{numeric vector \cr
A vector of estimates.}

\item{st_errs}{numeric vector \cr
A vector of estimated standard errors for each estimate.}

\item{k}{integer \cr
The rank of the estimate of interest. k = 1 returns the confidence interval
for the highest estimate, and k = length(estimates) returns the
confidence interval for the lowest estimate.}

\item{conf_level}{double in (0,1) \cr
The desired confidence level of the resulting interval. Note that
conditional and hybrid confidence intervals don't have nominal
coverage in the conventional sense.}
}
\description{
Returns an equal-tailed confidence interval with nominal coverage for the
k-th highest estimate in `estimates`.
}
